Have source of static_code in real .go file
For easier maintenance of the static_code, its content
has been moved to real .go files, which are excluded from
the normal builds with build tags.

The small tool called static_code_generator.go is used to generate
the necessary files for the builder package, by replacing the
header and wrapping the content in a variable.

The necessary commands to generate these files are also
integrated in the Makefile. This results in better maintenance
while the existing workflows are left untouched.