Optimize grammar and `-optimize-grammar` flag
The ast package is extended with an Optimize function,
which walks the given AST and applies several optimizations.

For the optimization, the following actions are taken:
* removal of unreferenced rules
* replace rule references with a copy of the referenced Rule, if the
  referenced rule it self has no references.
* resolve nested choice expressions
* resolve choice expressions with only one alternative
* resolve nested sequences expression
* resolve sequence expressions with only one element
* combine character class matcher and literal matcher, where possible
The resulting grammar is usually more memory consuming, but faster for parsing.

With the added flag `-optimize-grammar`, pigeon is instructed to generate
a parser with optimized grammar.