Correctly handle U+FFFD
U+FFFD, the replacement character, is a valid rune. However, it is also
used to signal EOF or invalid encoding when paired with 0 or 1 lengths.

This makes sure that both anyMatcher and charClassMatcher handle it
correctly, by only special-casing lengths 0 and 1.

The issue was partially fixed following #12 by not causing a parsing
error, but the matching was still incorrect. The test for issue 12
encoded invalid behavior (the replacement character was accepted even if
it was not in the character class) so it's removed.

Depending on the application, this might be a security issue.