Add state change expression (#{})
Implementation of the state handling proposal in #40.
AndCodeExpressions, NotCodeExpressions and ActionExpressions are now
readonly in regards to the state of the parser.
If the `-optimize` flag is used for the generated parser, the respective
save and restore operations for the parser state are removed.

Add InitState option which allows to initialize global "state" store with a value for
a given key.

Add goto example based of global state instead of GlobalStore.