Add Cloner interface
To allow pigeon to properly copy the global state, also for types
which are not passed by value (map, slice, chan) or structs
containing one of those values, the Cloner interface is added.
For types, which implement this interface, the correct cloning
is delegated to the respective type.

Add comment about "shallow" copy in global state

Fix return for InitState, Update documentation, make all