remove the emptyState optimization
In cloneState, the parser returns the emptyState map if the state is empty.
However, on restore it is set as the p.cur.state, so subsequent state code
blocks can add items to the emptyState, making it non-empty. To prevent
any issues because of this, I removed this small optimization.