Cache `want` in litMatcher
parseLitMatcher() was calling strconv.AppendQuote() each time, which my
profiler discovered was costing a measurably large amount of runtime
during parsing within dhall-golang.

This commit adds a cache field to the litMatcher struct so that we only
have to compute the `want` value once per litMatcher instead of once
each time we parse with the litMatcher.